Chapter12/requirements.txt
mysqlclient
click
blinker
redis
enum
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.1.1
Flask-WTF==0.14.2
flask-restful
flask-jwt-extended
flask-debugtoolbar
flask-admin
Flask-Caching
Flask-Babel==0.11.1
Flask-Mail
Faker==0.8.13
Flask-Assets
cssmin
jsmin
flask-bcrypt
flask-openid
flask-dance
celery
flask-celery-helper
selenium
coverage
